move postcode validator to validators folder
